Drop spannertest support and split session_test.go
